pyproject.toml
[tool.poetry]
name = "school-bot"
version = "0.1.0"
description = "A Telegram bot for school communication"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
telethon = "1.37.0"
httpx = "0.27.2"
sqlmodel = "0.0.22"
faststream = { extras = ["redis"], version = "0.5.30" }
beautifulsoup4 = "4.12.3"
pydantic = "2.9.2"
python-dotenv = "1.0.1"
pydantic-settings = "^2.6.1"
crawl4ai = "^0.3.731"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
soupsieve = "^2.6"
tabulate = "^0.9.0"
pyyaml = "^6.0.2"
sqlalchemy = "^2.0.36"
fastapi = {extras = ["all"], version = "^0.115.5"}
apscheduler = "^3.10.4"
anyio = "^4.6.2"
fast-depends = "^2.4.12"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
black = "24.10.0"
ruff = "0.7.3"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q"
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
